refactor: 共通ヘルパー関数をutilsに集約
