--- src/app/(main)/css/page.module.css
@@ -1,4 +1,0 @@
-.title {
-  font-size: 1.8rem;
-  margin-bottom: 1.5rem;
-}
