--- src/lib/actions.ts
@@ -1,20 +1,0 @@
-'use server';
-
-import { createSupabaseServerClient } from './supabase/server';
-
-/**
- * 認証済みユーザーのSupabaseクライアントとユーザー情報を取得するヘルパー関数
- * @returns {Promise<{supabase: SupabaseClient, user: User}>}
- * @throws ユーザーが認証されていない場合にエラーをスローする
- */
-export const getAuthenticatedClient = async () => {
-  const supabase = await createSupabaseServerClient();
-  const {
-    data: { user },
-  } = await supabase.auth.getUser();
-  if (!user) {
-    throw new Error('認証されていません。');
-  }
-
-  return { supabase, user };
-};
