+++ src/lib/utils.ts
@@ -1,6 +1,25 @@
-// src/lib/utils.ts
+'use server';
 
 import { type SupabaseClient, type User } from '@supabase/supabase-js';
+
+import { createSupabaseServerClient } from './supabase/server';
+
+/**
+ * 認証済みユーザーのSupabaseクライアントとユーザー情報を取得するヘルパー関数
+ * @returns {Promise<{supabase: SupabaseClient, user: User}>}
+ * @throws ユーザーが認証されていない場合にエラーをスローする
+ */
+export const getAuthenticatedClient = async () => {
+  const supabase = await createSupabaseServerClient();
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
+  if (!user) {
+    throw new Error('認証されていません。');
+  }
+
+  return { supabase, user };
+};
 
 /**
  * タイムライン系のコンポーネントで必要となるContextの値を生成するヘルパー関数
